Adding Cython

+++ tests/functional/test_decision_tree.py
@@ -8,8 +8,10 @@ def test_decision_tree_1():
     """Small repex to alert me to any issue."""
     from my_ml.model.decision_tree import DecisionTree
 
-    X: np.ndarray = np.array([[15, 0, 0], [12, 1, 0], [4, 1, 0], [8, 0, 1], [0, 0, 0]])
-    y: np.ndarray = np.array([0, 0, 1, 1, 1]).reshape(-1, 1)
+    X: np.ndarray = np.array(
+        [[15, 0, 0], [12, 1, 0], [4, 1, 0], [8, 0, 1], [0, 0, 0]], dtype=float
+    )
+    y: np.ndarray = np.array([0, 0, 1, 1, 1], dtype=float).reshape(-1, 1)
     dtree = DecisionTree()
     dtree.fit(X, y)
     X_test: np.ndarray = np.array([[0, 1, 0], [1, 0, 0]])
